Update 0029-PoWTargetTimespan.md
I dropped a term, but now it makes a bunch more sense to me with the `2(a-1)b`.  This is because when the PoSPoW interval is implemented in rfc0020, it uses a concept of `nTargetSpacing+nPoSSpacing`, which is this 'b+1' concept.  You can see it here, where at the end of the day my cross term has an 'a-1', but the +1 on the b term has been canceled out.