Update 0030-PoW-ASERT.md
Fixed some equations and changed paragraph 3 of the alternatives section.  This is a rather academic analysis of noise floors that does not relate to the following practical analysis of the proceeding paragraphs.  The definition of a proper figure of merit for the probability of the occurrence for a series of PoW blocks should not withhold progress on implementing this elegant upgrade.